+++ src/livestreamer/plugins/svtplay.py
@@ -1,13 +1,9 @@
 import re
 
 from livestreamer.plugin import Plugin
-from livestreamer.plugin.api import http, validate
+from livestreamer.plugin.api import StreamMapper, http, validate
 from livestreamer.stream import HLSStream, HDSStream
 
-STREAM_TYPES = {
-    "flash": HDSStream.parse_manifest,
-    "ios": HLSStream.parse_variant_playlist
-}
 
 _url_re = re.compile("""
     http(s)?://
@@ -28,7 +24,6 @@ _video_schema = validate.Schema(
                     "url": validate.text,
                     "playerType": validate.text
                 }],
-                validate.filter(lambda r: r["playerType"] in STREAM_TYPES)
             ),
         }
     },
@@ -42,21 +37,22 @@ class SVTPlay(Plugin):
     def can_handle_url(self, url):
         return _url_re.match(url)
 
+    def _create_streams(self, stream_type, parser, video):
+        try:
+            streams = parser(self.session, video["url"])
+            return streams.items()
+        except IOError as err:
+            self.logger.error("Failed to extract {0} streams: {1}",
+                              stream_type, err)
+
     def _get_streams(self):
         res = http.get(self.url, params=dict(output="json"))
         videos = http.json(res, schema=_video_schema)
-        streams = {}
-        for video in videos:
-            url = video["url"]
-            stream_type = video["playerType"]
-            parser = STREAM_TYPES[stream_type]
 
-            try:
-                streams.update(parser(self.session, url))
-            except IOError as err:
-                self.logger.error("Failed to extract {0} streams: {1}",
-                                  stream_type, err)
+        mapper = StreamMapper(cmp=lambda type, video: video["playerType"] == type)
+        mapper.map("ios", self._create_streams, "HLS", HLSStream.parse_variant_playlist)
+        mapper.map("flash", self._create_streams, "HDS", HDSStream.parse_manifest)
 
-        return streams
+        return mapper(videos)
 
 __plugin__ = SVTPlay