+++ src/livestreamer_cli/main.py
@@ -113,7 +113,7 @@ def iter_http_requests(server, player):
             continue
 
 
-def output_stream_http(plugin, streams):
+def output_stream_http(plugin, initial_streams):
     """Continuously output the stream over HTTP."""
 
     server = create_http_server()
@@ -126,7 +126,6 @@ def output_stream_http(plugin, streams):
     player = PlayerOutput(args.player, args=args.player_args,
                           filename=server.url,
                           quiet=not args.verbose_player)
-    stream_names = [resolve_stream_name(streams, s) for s in args.stream]
 
     try:
         console.logger.info("Starting player: {0}", args.player)
@@ -139,27 +138,23 @@ def output_stream_http(plugin, streams):
         user_agent = req.headers.get("User-Agent") or "unknown player"
         console.logger.info("Got HTTP request from {0}".format(user_agent))
 
-        stream = stream_fd = None
-        while not stream_fd:
-            if not player.running:
-                break
-
+        stream_fd = None
+        while not stream_fd and player.running:
             try:
-                streams = streams or fetch_streams(plugin)
-                for stream_name in stream_names:
-                    stream = streams.get(stream_name)
-                    if stream: break
-                else:
-                    stream = None
+                streams = initial_streams or fetch_streams(plugin)
+                initial_streams = None
 
+                for stream_name in (resolve_stream_name(streams, s) for s in args.stream):
+                    if stream_name in streams:
+                        stream = streams[stream_name]
+                        break
+                else:
+                    console.logger.info("Stream not available, will re-fetch "
+                                        "streams in 10 sec")
+                    sleep(10)
+                    continue
             except PluginError as err:
                 console.logger.error(u"Unable to fetch new streams: {0}", err)
-
-            if not stream:
-                console.logger.info("Stream not available, will re-fetch "
-                                    "streams in 10 sec")
-                streams = None
-                sleep(10)
                 continue
 
             try:
@@ -168,11 +163,9 @@ def output_stream_http(plugin, streams):
                 stream_fd, prebuffer = open_stream(stream)
             except StreamError as err:
                 console.logger.error("{0}", err)
-                stream = streams = None
-        else:
-            console.logger.debug("Writing stream to player")
-            read_stream(stream_fd, server, prebuffer)
 
+        console.logger.debug("Writing stream to player")
+        read_stream(stream_fd, server, prebuffer)
         server.close(True)
 
     player.close()
@@ -617,14 +610,14 @@ def setup_http_session():
     if args.https_proxy:
         livestreamer.set_option("https-proxy", args.https_proxy)
 
-    if args.http_cookies:
-        livestreamer.set_option("http-cookies", args.http_cookies)
+    if args.http_cookie:
+        livestreamer.set_option("http-cookies", dict(args.http_cookies))
 
-    if args.http_headers:
-        livestreamer.set_option("http-headers", args.http_headers)
+    if args.http_header:
+        livestreamer.set_option("http-headers", dict(args.http_header))
 
-    if args.http_query_params:
-        livestreamer.set_option("http-query-params", args.http_query_params)
+    if args.http_query_param:
+        livestreamer.set_option("http-query-params", dict(args.http_query_param))
 
     if args.http_ignore_env:
         livestreamer.set_option("http-trust-env", False)
@@ -640,6 +633,22 @@ def setup_http_session():
 
     if args.http_timeout:
         livestreamer.set_option("http-timeout", args.http_timeout)
+
+    if args.http_cookies:
+        console.logger.warning("The option --http-cookies is deprecated since "
+                               "version 1.11.0, use --http-cookie instead.")
+        livestreamer.set_option("http-cookies", args.http_cookies)
+
+    if args.http_headers:
+        console.logger.warning("The option --http-headers is deprecated since "
+                               "version 1.11.0, use --http-header instead.")
+        livestreamer.set_option("http-headers", args.http_headers)
+
+    if args.http_query_params:
+        console.logger.warning("The option --http-query-params is deprecated since "
+                               "version 1.11.0, use --http-query-param instead.")
+        livestreamer.set_option("http-query-params", args.http_query_params)
+
 
 def setup_plugins():
     """Loads any additional plugins."""